Initialize Repositories, Controllers, Services, and Utilities

+++ src/main/java/com/letter_loom/entities/User.java
@@ -19,8 +19,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id;
-
+    private Long id;
 
     @Column(name = "username")
     private String username;
